requirements.txt
Flask
google-cloud-aiplatform>=1.38.0
python-dotenv
torch>=2.5.1
torchvision>=0.20.1
diffusers>=0.32.1
transformers>=4.47.1
einops>=0.7.0
accelerate>=1.2.1
